move parse to common location:
